requirements.txt
accelerate==1.3.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
datasets==2.16.1
dill==0.3.8
filelock==3.17.0
filetype==1.2.0
frozenlist==1.5.0
fsspec
ftfy==6.3.1
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==10.4.0
propcache==0.2.1
psutil==6.1.1
pyarrow<14.0.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
setuptools==75.8.0
six==1.17.0
surya-ocr==0.4.5
sympy==1.13.3
tabulate==0.9.0
tokenizers==0.15.2
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.36.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.45.1
xxhash==3.5.0
yarl==1.18.3
